Added a directory for images and updated the recipepage for the smashburger.
The images directory will hold the images for all the images in the project. Currently
it only contains an image of the smashburger, but will get more. Secondly I updated the
recipe page for the smashburger to be finalized, not accounting for the css that I will
add onto it in a future part of the Odin project of course.