Added another image in images for the new recipe that I now link to in the index page. The added recipe is of a MousseCake which I found from Camilla Hamid.
The page for the MousseCake has about the same structure as the smashburger. The keydifference
is that the mousscake has seperate parts that have to be done at differnet stages,
that is reflected in both the structure of the webpage and the steps in the steps part of the webpage.